Added valid time to json output of ota
